Tweaks and new Global Colors Files
