+++ lib/practiceExam.dart
@@ -1,5 +1,6 @@
 //import 'package:getwidget/getwidget.dart';
 import 'package:flutter/material.dart';
+import 'package:lsat_prep_app_2/colorSchemes.dart';
 import 'package:lsat_prep_app_2/Profile.dart';
 import 'package:lsat_prep_app_2/homePage.dart';
 import 'package:lsat_prep_app_2/profile_page.dart';
@@ -56,7 +57,7 @@ class _pTestState extends State<pTestPage> {
           if (index == 0){
             Navigator.push(
                 context, MaterialPageRoute(builder: (context) {
-              return const homeDesign();
+              return const homePage();
             }));
           }
           if (index == 1){
