+++ lib/settings.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:lsat_prep_app_2/colorSchemes.dart';
 import 'package:lsat_prep_app_2/homePage.dart';
 import 'package:lsat_prep_app_2/Profile.dart';
 import 'package:lsat_prep_app_2/profile_page.dart';
@@ -84,7 +85,7 @@ class settings extends StatelessWidget {
             if (index == 0){
               Navigator.push(
                   context, MaterialPageRoute(builder: (context) {
-                return const homeDesign();
+                return const homePage();
               }));
             }
             if (index == 1){
